+++ test/dummy/rails61_webpacker6/app/models/ship.rb
@@ -1,0 +1,3 @@
+class Ship < ApplicationRecord
+  has_many :members
+end