+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Initial support for ActionText
 * Generator for setting up ActionText. Before Super can work with ActionText,
   you'll first need to set up Webpacker and ActionText on your application
+* Initial support for Webpacker 6
 
 
 ## [0.0.8] - 2020-12-27